requirements.txt
streamlit==1.51.0
seaborn
plotly
scikit-learn
openpyxl
pandas
numpy
matplotlib
openai
python-dotenv